backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
requests
httpx
billboard.py
spotipy
numpy
scikit-learn
soundcharts
google-generativeai>=0.8.0
google-auth
google-auth-oauthlib
google-auth-httplib2
pyjwt
python-jose[cryptography]
pytrends
yfinance
lyricsgenius
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
scipy
textblob
nltk
pandas
matplotlib
seaborn 